Adding a faster version of write_stripped_pdb (takes 61% of the time). Adding an option to pass keep_hetatm_lines for Tina's run. Adding a progress counter for the file creation stage. Adding (kind of hacky) checkpointing to avoid re-writing files if all files exist. Disabled at present.
Todo:
1. Option to not call write_stripped_pdb;
2. Enable and test re-writing checkpointing;
3. Replace write_stripped_pdb.

Bugfix: job_output_directory is now created if it does not exist.

+++ create_ddg_monomer_interface_run.py
@@ -14,17 +14,33 @@ import re
 import shutil
 import importlib
 from klab.cluster_template.write_run_file import process as write_run_file
+from klab import colortext
+from klab.fs.fsio import write_file
 
 job_output_directory = 'job_output'
 
+
+time_spent_in_write_stripped_pdb = 0.0
+
 def write_stripped_pdb(new_file_location, file_contents):
+    global time_spent_in_write_stripped_pdb
+    t1 = time.time()
     if isinstance(file_contents, basestring):
         file_contents = file_contents.split('\n')
     with open(new_file_location, 'w') as f:
         for line in file_contents:
             if line.startswith('ATOM'):
                 f.write(line + '\n')
-    
+    time_spent_in_write_stripped_pdb += time.time() - t1
+
+
+def write_stripped_pdb(new_file_location, file_contents):
+    global time_spent_in_write_stripped_pdb
+    t1 = time.time()
+    write_file(new_file_location, '\n'.join([l for l in file_contents.split('\n') if l.startswith('ATOM')]))
+    time_spent_in_write_stripped_pdb += time.time() - t1
+
+
 if __name__ == '__main__':
     assert( len(sys.argv) > 1 )
     cfg = importlib.import_module(sys.argv[1], package=None)
@@ -37,16 +53,26 @@ if __name__ == '__main__':
     rosetta_scripts_path = settings['local_rosetta_installation_path'] + '/source/bin/' + 'rosetta_scripts' + settings['local_rosetta_binary_type']
     ppi_api = get_interface_with_config_file(rosetta_scripts_path = rosetta_scripts_path, rosetta_database_path = '/home/kyleb/rosetta/working_branches/alascan/database')
 
+    #ppi_api.destroy_prediction_set('ran_gsp1_ddg-monomer-16_1')
+
     if not ppi_api.prediction_set_exists(prediction_set_id):
         print 'Creating new prediction set:', prediction_set_id
         ppi_api.add_prediction_set(prediction_set_id, halted = True, priority = 7, allow_existing_prediction_set = False, description = cfg.prediction_set_description)
 
+    # Read the keep_hetatm_lines optional setting
+    keep_hetatm_lines = False
+    try: keep_hetatm_lines = cfg.keep_hetatm_lines
+    except: colortext.warning('Note: keep_hetatm_lines is not specified in {0}. Defaulting to {1}.'.format(sys.argv[1], keep_hetatm_lines))
+
     # Populate the prediction set with jobs from a (tagged subset of a) user dataset
-    ppi_api.add_prediction_run(prediction_set_id, cfg.user_dataset_name, tagged_subset = cfg.tagged_subset, extra_rosetta_command_flags = '-ignore_zero_occupancy false -ignore_unrecognized_res', show_full_errors = True)
     print 'Created PredictionSet:', prediction_set_id
+    ppi_api.add_prediction_run(prediction_set_id, cfg.user_dataset_name, keep_hetatm_lines = keep_hetatm_lines, tagged_subset = cfg.tagged_subset, extra_rosetta_command_flags = '-ignore_zero_occupancy false -ignore_unrecognized_res', show_full_errors = True)
 
     existing_job = False
     end_job_name  = '%s_%s' % (getpass.getuser(), prediction_set_id)
+    if not os.path.exists(job_output_directory):
+        os.makedirs(job_output_directory)
+
     for d in os.listdir(job_output_directory):
         if os.path.isdir(os.path.join(job_output_directory, d)) and end_job_name in d:
             print 'Found existing job:', d
@@ -93,7 +119,17 @@ if __name__ == '__main__':
     app_name = 'minimize_with_cst'
     settings['appname'] = app_name
 
+    # Progress counter setup
+    colortext.message('Creating input data for %d predictions.' % (len(prediction_ids)))
+    count, records_per_dot = 0, 50
+    print("|" + ("*" * (int(len(prediction_ids)/records_per_dot)-2)) + "|")
+
+    num_stripped_files_written = 0.0
     for prediction_id in prediction_ids:
+        # Progress counter
+        count += 1
+        if count % records_per_dot == 0: colortext.write(".", "cyan", flush = True)
+
         # Check if job already ran
         prediction_id_dir = os.path.join(output_dir, str(prediction_id))
         if existing_job:
@@ -118,19 +154,30 @@ if __name__ == '__main__':
         substitution_parameters = json.loads(job_details['JSONParameters'])
 
         job_data_dir = os.path.join(output_data_dir, str(prediction_id))
-        if os.path.isdir(job_data_dir):
-            shutil.rmtree(job_data_dir)
-        os.makedirs(job_data_dir)
+        # Allow us to resume from an interrupted setup
+        all_files_exist = os.path.exists(job_data_dir) and os.path.exists(os.path.join(job_data_dir, '.ready'))
+        all_files_exist = False # todo: remove
+        if not all_files_exist:
+            if os.path.isdir(job_data_dir):
+                shutil.rmtree(job_data_dir)
+            os.makedirs(job_data_dir)
 
         files_dict = {} # Maps name to filepath position
         for file_name, file_contents in file_tuples:
             new_file_location = os.path.join(job_data_dir, file_name)
-            if '.pdb' in file_name:
-                write_stripped_pdb(new_file_location, file_contents)
-            else:
-                with open(new_file_location, 'w') as f:
-                    f.write(file_contents)
+            if not all_files_exist:
+                if '.pdb' in file_name:
+                    write_stripped_pdb(new_file_location, file_contents)
+                    num_stripped_files_written += 1.0
+                    if num_stripped_files_written > 100:
+                        break # todo remove
+                else:
+                    with open(new_file_location, 'w') as f:
+                        f.write(file_contents)
             files_dict[file_name] = os.path.relpath(new_file_location, settings['output_dir'])
+
+        if num_stripped_files_written > 100:
+            break # todo remove
 
         # Figure out input fi
             
@@ -139,9 +186,11 @@ if __name__ == '__main__':
         }
         job_dict[prediction_id] = argdict
 
+    print('')
     if len(job_dict) > 0:
         write_run_file(settings, database_run = False, job_dict = job_dict)
         print 'Job files written to directory:', os.path.abspath(output_dir)
     else:
         print 'No tasks to process, not writing job files'
 
+    print('time in time_spent_in_write_stripped_pdb is {0}s: {1}s per file ({2} files).'.format(time_spent_in_write_stripped_pdb, time_spent_in_write_stripped_pdb / num_stripped_files_written, num_stripped_files_written))
